Tentative implementation for CodeCommit support [ci fast]
