+++ modules/nextflow/src/main/resources/META-INF/extensions.idx
@@ -19,3 +19,4 @@ nextflow.trace.DefaultObserverFactory
 nextflow.util.DefaultSerializers
 nextflow.secret.LocalSecretsProvider
 nextflow.cache.DefaultCacheFactory
+nextflow.scm.RepositoryFactory
